Divide o Quicksort em iterativo e recursivo
